// interfaces/Items/IItemSellPrice.ts
export default interface IItemSellPrice {
  item_id: number
  item_stock_id: number
  item_sell_price_id: number
  price_category_name: string
  purchase_price: number
  cost: number
  profit: number
  item_name: string
  warehouse_name: string
  item_code: string
  unit_name: string
  sell_price: number
  category_name: string
  sub_category_name: string
  default_price: number
  quantity: number
}